+++ src/components/choices.jsx
